Rename dash stubs to com.brentvatne.exoplayer.dashstubs
