+++ android/src/main/java/com/brentvatne/exoplayer/ReactExoplayerView.java
@@ -64,13 +64,13 @@ import androidx.media3.datasource.HttpDataSource;
 import androidx.media3.exoplayer.DefaultLoadControl;
 import androidx.media3.exoplayer.DefaultRenderersFactory;
 import androidx.media3.exoplayer.ExoPlayer;
-import androidx.media3.exoplayer.dash.DashMediaSource;
-import androidx.media3.exoplayer.dash.DashUtil;
-import androidx.media3.exoplayer.dash.DefaultDashChunkSource;
-import androidx.media3.exoplayer.dash.manifest.AdaptationSet;
-import androidx.media3.exoplayer.dash.manifest.DashManifest;
-import androidx.media3.exoplayer.dash.manifest.Period;
-import androidx.media3.exoplayer.dash.manifest.Representation;
+import com.brentvatne.exoplayer.dashstubs.DashMediaSource;
+import com.brentvatne.exoplayer.dashstubs.DashUtil;
+import com.brentvatne.exoplayer.dashstubs.DefaultDashChunkSource;
+import com.brentvatne.exoplayer.dashstubs.manifest.AdaptationSet;
+import com.brentvatne.exoplayer.dashstubs.manifest.DashManifest;
+import com.brentvatne.exoplayer.dashstubs.manifest.Period;
+import com.brentvatne.exoplayer.dashstubs.manifest.Representation;
 import androidx.media3.exoplayer.drm.DefaultDrmSessionManager;
 import androidx.media3.exoplayer.drm.DefaultDrmSessionManagerProvider;
 import androidx.media3.exoplayer.drm.DrmSessionEventListener;
