+++ android/src/main/java/com/brentvatne/exoplayer/dashstubs/DashUtil.java
@@ -1,9 +1,10 @@
-package androidx.media3.exoplayer.dash;
+package com.brentvatne.exoplayer.dashstubs;
 
 import android.net.Uri;
 
 import androidx.media3.datasource.DataSource;
-import androidx.media3.exoplayer.dash.manifest.DashManifest;
+
+import com.brentvatne.exoplayer.dashstubs.manifest.DashManifest;
 
 public class DashUtil {
     public static DashManifest loadManifest(DataSource ds, Uri uri) {
