+++ android/src/main/java/com/brentvatne/exoplayer/dashstubs/manifest/Representation.java
@@ -1,4 +1,4 @@
-package androidx.media3.exoplayer.dash.manifest;
+package com.brentvatne.exoplayer.dashstubs.manifest;
 
 import androidx.media3.common.Format;
